backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
google-api-python-client
python-dotenv
pillow
google-genai
